Required components and initialization
Added RequireComponent to FeedPlayer to require both MixingAudioPlayer and Session components as they will always be need.

Using the Unity Editor's lifecycle, we also made the binding happen automatically. As soon the FeedPlayer component is added, the references to MixingAudioPlayer and Session are initialized. This makes using the SDK much smoother on the editor side. The user of the SDK now only has to add FeedPlayer and all else is handled by FeedFM.

The binding of the two required scripts will be done during compile time as oppose to the previous SDK which was do during runtime. This change has removed the memory fragmentation that exists in the previous SDK.

DisallowMultipleComponent has been added to FeedPlayer, MixingAudioPlayer, and Session as only one of each component is required to create a working set.

For legacy compatibility, we move FeedPlayer up twice to maintain the order below
- FeedPlayer
- MixingAudioPlayer
- Session
This will allow for a smoother migration for users of the previous SDK who have code that depends on the order of initialization.