Added a read only attribute
Added a custom read only attribute to provide a visual aid to consumers of the SDK. Fields that are read only show that the intent of that field is not to be edited. Fields with this attribute are useful for viewing information at a glance from the editor without the risk of being modified by accident and removes the need to log to view the values.